feat(execution): plumb relevance threshold through eval and binarize >= T_r Extend polytris manifests with EXEC RELEVANCE_THRESHOLDS and carry relevance_threshold through prune breakdown/timelimit, throughput and accuracy manifests, tradeoff columns, and compare charts (tooltips, filters, backward-compatible NA columns). Binarize classification bitmaps at int(T_r * 255) using >= in stage_compress and p030_exec_compress for boundary-consistent cutoff behavior. Align p026/p027 original-tile retention counts with p022 by thresholding relevancy scores at int(T_r * 255) instead of a fixed 128. Add evaluation/p017_mistrack_variance.py to write paper LaTeX macros for mistrack variance and standard deviation from cached track_rates.

+++ execution/stage_compress.py
@@ -99,7 +99,7 @@ def _compress_video(
         hex_data: str = frame_result['classification_hex']
         cls_size: tuple[int, int] = frame_result['classification_size']
         bitmap = np.frombuffer(bytes.fromhex(hex_data), dtype=np.uint8).reshape(cls_size)
-        bitmap = (bitmap > int(threshold * 255)).astype(np.uint8)
+        bitmap = (bitmap > (threshold * 255)).astype(np.uint8)
         assert dtypes.is_bitmap(bitmap), bitmap.shape
         polyominoes_stacks[array_idx] = group_tiles(bitmap, TILEPADDING_MAPS[tilepadding])
